Merged Nunit Support into Purple
Had to remove FiddlerProxy.cs and Sazformat.cs to get tests to compile
and run.  BrowserMobProxy.cs is included and removed duplicates in Obj
file from the ProtoTest.Golem project to get tests to compile and run.
It will take some work to get purple and Nunit to play nice together
officially.