Added the RightClick functionality. Added a new constructor which receives a text for the element's name and an AutomationElement.
I created a custom element in the tests called LQP_PurpleButton, this Element is created from an AutomationElement instead of a purplePath because the purplePath I was working with was not the kind of paths we usually use.
Sometimes we found AutomationElements from another AutomationElement, and there is not a straight forward way to create a PurpleElementBase, or any of the inherited elements, from the AutomatedElement. This is a first approach to that.