Removed Attach to Process Functionality
This was causing some problems with tests attempting to attach to a
running process that was being killed.  The test would attach to a
process, then the process would be killed, and the test would stall.
For now all tests start a new process before running the test.  Test
should be able to run faster if we can attach to existing processes.
That will take some additional planning.